# Refactor: Optimize maplayers