--- a/src/game/client/components/maplayers.h
+++ b/src/game/client/components/maplayers.h
@@ -125,11 +125,11 @@ class CMapLayers : public CComponent
 		bool m_IsTextured;
 	};
 	std::vector<SQuadLayerVisuals *> m_vpQuadLayerVisuals;
+	std::vector<std::vector<int>> m_vvLayerCount;
+	int m_GameGroup;
 
 	virtual CCamera *GetCurCamera();
 	virtual const char *LoadingTitle() const;
-
-	void LayersOfGroupCount(CMapItemGroup *pGroup, int &TileLayerCount, int &QuadLayerCount, bool &PassedGameLayer);
 
 public:
 	enum